moved files and deleted package. Fixed output path, as in review wanted. Replaced windows paths with normal /
