+++ src/main/java/spoon/refactoring/MethodInvocationSearch.java
@@ -3,7 +3,7 @@
  *
  * Spoon is available either under the terms of the MIT License (see LICENSE-MIT.txt) of the Cecill-C License (see LICENSE-CECILL-C.txt). You as the user are entitled to choose the terms under which to adopt Spoon.
  */
-package spoon.refactoring.invocation;
+package spoon.refactoring;
 
 import java.util.ArrayList;
 import java.util.Arrays;
